Update Package-development-cycle.rmd
Added and removed a few commas. Made the lists a bit more parallel in structure by making each element a sentence ending in a period. I made the use of commas consistent for lists of three or more elements by placing a comma be for the conjunction `and'. I removed the commas before the conjunctions when the list only contained two elements.