Update Special-environments.rmd
I think x (in the case of curve2) and .it (in the case of with_conn) should be assigned in the newly constructed environment, not in the parent.frame(). This way they are properly enclosed.

Also looks like you had a typo in eval'ing the with_conn code in the context of env when it should have been env2.

Great book!