Update OO-essentials.rmd
Expand the introduction slightly to contrast with functional programming - recommend "The Object-Oriented Thought Process" as a text for those who might want to read up on this in more detail (not sure if you've read it, Hadley - feel free to replace it if not/if you know something better). I'd include the upcoming OOP in R book but it's, well, upcoming.