Minor typo suggestion
On the rendered website, the console output for a code chunk (lines 439-445) should be

#> a + b + x

But is instead

#> 1 + 2 + mpg

I am not sure how to fix this, but wanted to point it out. Appears to be a quirk in how knitr converts .Rmd to .md, but I do not know enough about knitr internals to know for sure.