align code to framing text
The text asks the reader to think about this for a first and second execution, so probably we don't want to a) call the function once and b) show the output.

As a note, this is almost identical to the example in the section "A fresh start" in "Functions". This is probably fine, but for some readers going straight through the duplication could be a little irksome.