simplify (fix?) code
Right now the rendered version looks like this:

"""
x <- 10
lockBinding(as.name("x"), globalenv())
#> Error: no binding for "x"
x <- 15
rm(x)
"""

I'm not sure why this is. On my machine at least, it seems that "as.name()" is not necessary for this snippet to work - but I also can't replicate the "no binding" error. I defer to your judgement if this change is not good.