Update Documenting-functions.rmd
The terminology RC is used in an earlier chapter about OO programming in R. As a reader I get confused about reading about R5 classes all of the sudden.

The link, in the link section to the left, should also be updated to RC from R5 if this proposal is approved. Cannot find where to do this update in these pages.