lapply is NOT equivalent to for-loop
Lines 551-578. Somehow the lapply function is not equivalent to the for loop in this case. You know better, but the two outputs of the list of closures adders and adders2 is the SAME even in your book.
Perhaps, the reason is that lapply calls .Internal (maybe not, you know better), but there is something fundamentally different between simple for-loop and lapply in this case. Therefore, the lazy evaluation is to be demonstrated only with for-loops and not lapply function.
Thanks!