HTML different than XML
HTML and XML are quite different and that is deliberate by the recommendation, i.e. spec, for XML. 

Note: The book example is invalid html (for any version). An <image> tag must have an alt attribute and there has to be a <title> tag in the header.

I would remove any reference to XML, it really doesn't add to the understanding of HTML. The second paragraph I've added is an alternative for the first, but with no reference XML.

Here is an example completely valid HTML 5 that is not correct, i.e. not well formed, nor similar to XML.

<!DOCTYPE html >
<title>test</title>
<h1 id=first>A heading</h1>
<p>Some text &amp; <b>some bold text.</b>

<img src=mying.png width=100 height=100 alt="an image"/>

Notice that there is no root tag at all, i.e. no <body> tag nor a <html> tag. Notice the lack of quotes on some attribute values. Notice there is no closing </p> tag. 

Any HTML parser will create and AST of this page that includes html, body, etc. elements. Try this example in Chrome (or your favorite browser) and  use the developer tools to see the AST. That's how the W3C DOM says a parser should work.

Also it easy to make working example with html, just add a DOCTYPE and a <title> and alt to the example. I think that whenever possible examples should work and be correct. As is the example works but is not correct.

The comment about HTML and XML being similar isn't correct. HTML and "X"HTML are similar but for all practical purposes XHTML is deprecated (I don't know if it formally is, but in practice it shouldn't be used in new projects).

In fact HTML and XML are deliberately different. One of the explicit goals of the XML Recommendation was "Terseness of markup is not a goal." One of the reasons Berners-Lee specified a dialect of SGML that didn't need things like an <html> tag was because at the time it was almost always authored by hand.

Sorry but I done a lot of HTML and XML at the plumbing level (for Microsoft's msdn.com site) and "trivial" details of the HTML recommendation often caused havoc because no two browsers handled "out of spec" HTML differently.