quantum??
I don't know how the term quantum fits here... maybe I a missing something. Heisenberg uncertainty principle? 

I think the key concept being presented here is that promises are opaque object and that from R you cannot determine if an object is a promise or not, nor to determine a promise's environment. But that much detail isn't really needed.

All the reader needs to know is that an expression and an environment is all you need to evaluate an expression at a later time and that the capabilities of promises are very similar to quosures.