Changing `h` to `b` in composite functions
The `h` value in the `midpoint_composite` and `trapezoid_composite` functions appears to the hold the value of the base of the rectangle and trapezoid respectively. Given the name (`h`), one would assume that the variable represents the height. This change updates the variable name to better reflect what it represents in the equation.