Fixing some typos in chapter 19
Most of these are simple typo fixes, but I did run into a bit of a problem following some of the conversation around `expr`, `enexpr`, and `exprs`. I tried to fill in what I thought was missing in some of those cases, but there's a very good chance that what I filled in is wrong, so feel free to discard as you see fit.