Group the alike.
The previous section interrupts the flow between the first time S3, S4, and other OO systems are introduced and this material -- were one section is dedicated to each of those OO systems introduced earlier.

My previous comment suggested a change that is consistent with the change I suggest here. That is, if the difference between base objects and OO is moved up, then the introduction to each OO system and their respective sections will be left in closer contact to one another.