Write for programmers of functions more than for users of functions
This is a selfish suggestion -- I want to learn how S3 works to write functions with it. But I'll try to sell my point. 

For users, S3 just works. If you write for the users of S3 functions, I guess the reaction you can expect is: "Good to know, but this won't change anything of what I already do." But if you write for the programmers of S3 functions, then the reaction could be totally different: "Wow, I'll change my approach to programming right now!".