Avoid confusion with S3 styles
In the context of S3 styles, the `list()` in `structure(list() ... ` confuses me. It suggests that all S3 objects are built on top of a list. This gets confusing because the scalar-style does build on top of lists.

Another source of confusion is that the list is empty. From the viewpoint of the data structure, an empty list is indeed simple. But from the viewpoint of a struggling reader (and I've read this topic multiple times), an empty list is not the data structure I deal most commonly with -- so it takes me an extra effort to understant if the data _has_ to be an empty list, instead of any data.