Some typos near the end of Quotation.Rmd
I assign the copyright of this contribution to Hadley Wickham.

PS: I have made a couple of small pull requests with typos in this chapter now. If it is preferable that I rather collect such corrections into one slightly larger pull request, please let me know, and I'll adjust in the future.