rename use of lang_standardise to call_standardise
Section 21.4.4 (Expressions::Data_structures::Calls)

The text in this section referred to `rlang::lang_standardise` in two
locations. This function has been deprecated in `rlang `and renamed
`call_standardise`.

The exercises at the end of the 'Data structures' subsection use the function
call_standardise, so I updated the text to refer to call_standardise not
lang_standardise.

I assign the copyright of this contribution to Hadley Wickham