Change reference link to cppreference
The C++ community seems to have converged to using cppreference as the "go to" place for online documentation and reference of the standard.  A key advantage is that cppreference is a community wiki -- an open platform as opposed to a closed one -- and generally have higher quality content. Another is that it doesn't have advertisements.

See the discussions here:

- https://stackoverflow.com/questions/6520052/whats-wrong-with-cplusplus-com
- https://meta.stackexchange.com/questions/194788/links-being-changed-to-cppreference-com

While the answers do not completely settle the issue, I find the arguments for preferring cppreference over cplusplus convincing.