alter the use of 'bootstrap'
As a [bootstrapper in a previous life](https://scholar.google.com/scholar?hl=en&as_sdt=0%2C37&q=wh+beasley+bootstrap&btnG=), this chapter is offensive.

I'm kidding around, but I do think the term 'bootstrap' is a stretch, especially in the 'Quoting and unquoting' section.  In this first section, it doesn't meet most definitions of bootstrap because it lacks (a) a test-stat/function applied to the sampling frame and (b) a distribution of that function's values.  The `boot_lm()`s (in the 'The evaluation environment' section) now apply a function to a sampled draw, but still don't produce a distribution of the test stat

To me, the existing `bootstrap()` is a specialized sampling function with no direct connection to any statistical inference (about standard errors, etc).  Consider renaming `bootstrap()` to something like `sample_rows()`/`draw_rows()`/`pluck_rows()`.

I think this example is well-suited for your larger purpose of demonstrating expressions & environments and I from it (even though I've read [your nse vignette](http://adv-r.had.co.nz/Computing-on-the-language.html) 10+ times over the psat 2 years).  And I'm not advocating that you introduce a full sampling distribution in to the function.  Just advocating that a different term is used.  Your books are so popular and influential to so many fields, I don't want non-statisticians from getting the wrong 1st impression of the technique.