Fixes to typos and possible typos in OOP section (#1426)
* Remove possible extract "from"

This sentence seems to make more sense with "from" removed.

* Insert possible missing "not"

Now "it may be challenging to fully understand if you have not done object oriented programming before." Although, I don't know, maybe it's easier to learn R's OOP if you've never learned it in another language? In which case, stet.

* Change "accidental" to "accidentally"

* Change "argument" to "arguments"

* Insert a comma to make a sentence easier to read

Without the comma, this sentence was really catching me up. The comma's not required, but it might help some readers.

* Change "an" to "a"

* Change "a" to "an"

In other spots, the text uses "an" before "R6", so this change is to be consistent with that.

* Change "the" to "then"

* Insert missing "how"

* Change "a" to "an"

Other spots use "an" before "S3", so this is to be consistent with that.

* Insert a possible missing "and"

I'm not sure about this one, but it seems to make more sense if you add the "and".

* Change "it's" to "its"

* Insert missing "that"

* Insert missing "a"

* Insert missing "that"

* Insert missing "that"

* Insert possible missing "include"

I wasn't sure about this one, but it seemed to make more sense with "include" added

* Replace "." with "?"

* Insert missing "that"

* Insert hyphen in "pseudo-class"

This is to make it consistent with other parts of the text.

* Delete repeated "same"

* Insert possible missing "it"s

This was easier for me to read with the "it"s, although it's probably fine as is.