Informando sobre o operador `:::`
I think it would be good, in this section, to inform the role of the `:::` operator. The operator is used more explicitly in Section 10.3.3. Informing in the Section where the S3 object-oriented system is discussed would, in my opinion, be a good opportunity to make clear to the reader the usefulness of the `:::` operator.

Best regards.